Moving new blocks to blocks folder
