+++ packages/js/product-editor/src/blocks/conditional/block.json
