+++ packages/js/product-editor/src/blocks/inventory-email/block.json
