+++ packages/js/product-editor/src/blocks/inventory-email/edit.tsx
