+++ packages/js/product-editor/src/blocks/inventory-email/index.ts
