+++ packages/js/product-editor/src/blocks/inventory-email/style.scss
