+++ packages/js/product-editor/src/components/editor/init-blocks.ts
@@ -23,8 +23,8 @@ import { init as initCollapsible } from '../collapsible-block';
 import { init as initScheduleSale } from '../../blocks/schedule-sale';
 import { init as initTrackInventory } from '../../blocks/track-inventory';
 import { init as initSku } from '../../blocks/inventory-sku';
-import { init as initConditional } from '../conditional-block';
-import { init as initLowStockQty } from '../inventory-email-block';
+import { init as initConditional } from '../../blocks/conditional';
+import { init as initLowStockQty } from '../../blocks/inventory-email';
 
 export const initBlocks = () => {
 	const coreBlocks = __experimentalGetCoreBlocks();
